feat : reestructuracion Backend + genre_service + tests
