+++ Backend API/src/services/song_service.py
@@ -165,25 +165,3 @@ async def create_song(name: str, artist: str, genre: Genre, photo: str, file) ->
     file_id = gridFsSong.put(
         file, name=name, artist=artist,duration=duration ,genre=str(genre.value), photo=photo)
 
-
-def get_genres() -> json:
-    """ Returns a json with all the available genres"
-
-    Parameters
-    ----------
-
-    Raises
-    -------
-
-    Returns
-    -------
-        Json { GenreEnum : 'genre'}
-    """
-
-    # Obtener todas las propiedades de la clase Genre
-    genre_properties = [(g.name, g.value) for g in Genre]
-
-    genre_dict = dict(genre_properties)
-    genre_json = json.dumps(genre_dict)
-
-    return genre_json
